cli/pyproject.toml
[tool.poetry]
name = "synthlang-cli"
version = "0.1.0"
description = "Command-line interface for the SynthLang framework using DSPy"
authors = ["SynthLang Team"]
readme = "README.md"
packages = [{include = "synthlang"}]

[tool.poetry.dependencies]
python = "^3.8"
dspy-ai = "^2.0.0"
click = "^8.1.7"
python-dotenv = "^1.0.0"
pydantic = "^2.5.2"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
synthlang = "synthlang.cli:main"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
